Resolve relative symlinks
The Linux kernel code-base can contain relative symlinks,
 like for example:
  include/dt-bindings/input/linux-event-codes.h -> ../../uapi/linux/input-event-codes.h

Cscope complains about a file not found when it looks for:
  ../../uapi/linux/input-event-codes.h

This patch adds a small pre-processing on the files list generation to ensure
that all files are reported using their actual full (relative) path.
This avoid cscope complaining and it has the additional benefit to get a full
path for each cscope symbol.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>